requirements.txt
pandas
langchain
langchain-openai
langchain-groq
langchain-community
python-dotenv
streamlit
